修正添加附件的问题
在后续判断中，magic num 5 = sizeof(char) +
sizeof(int)，如果变更魔数会导致和旧的文件格式不兼容，无法读取附件信息。

+++ libythtbbs/binaryattach.c
@@ -9,7 +9,7 @@ static int
 appendonebinaryattach(char *filename, char *attachname, char *attachfname)
 {
 	FILE *fp;
-	size_t size, origsize;
+	int size, origsize;
 	struct mmapfile mf = { ptr:NULL };
 	char ch = 0;
